backend/requirements.txt
google-genai
python-dotenv
google-cloud-bigquery
pandas
db-dtypes
fastapi
uvicorn[standard]
